requirements.txt
aiohttp==3.12.15
django-cors-headers==4.9.0
azure-storage-blob==12.26.0
boto3==1.40.18
cachetools==6.2.0
celery==5.5.3
coverage==7.10.7
Django==5.2.6
django-crontab==0.7.1
django-redis==6.0.0
django-tenants==3.9.0
django-tenant-schemas==2.0.0
djangorestframework==3.16.1
gcsfs==2025.9.0
gevent==25.9.1
google-cloud-storage==3.4.0
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
jsonmerge==1.9.2
msgspec==0.19.0
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.3
PyJWT==2.10.1
pydomo==0.3.0.13
psycopg2-binary==2.9.10
python-calamine==0.5.3
python-dateutil==2.9.0.post0
querystring-parser==1.2.4
redis==6.4.0
requests==2.32.5
s3fs==2025.9.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
XlsxWriter==3.2.9